accessible/tests/mochitest/actions/test_keys_menu.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        title="Accessible XUL access keys and shortcut keys tests">

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />

  <script type="application/javascript"
          src="../common.js" />
  <script type="application/javascript"
          src="../events.js" />

  <script type="application/javascript">
  <![CDATA[
    function openMenu(aMenuID, aMenuitemID)
    {
      this.menuNode = getNode(aMenuID),
      this.menuitemNode = getNode(aMenuitemID),

      this.eventSeq = [
        new invokerChecker(EVENT_FOCUS, this.menuNode)
      ];

      this.invoke = function openMenu_invoke()
      {
        // Show menu.
        this.menuNode.open = true;
      }

      this.finalCheck = function openMenu_finalCheck()
      {
        var menu = getAccessible(aMenuID);
        is(menu.keyboardShortcut, (MAC ? "u" : "Alt+u"),
           "Wrong accesskey on " + prettyName(this.menuitemNode));

        var menuitem = getAccessible(aMenuitemID);
        is(menuitem.keyboardShortcut, "p",
           "Wrong accesskey on " + prettyName(this.menuitemNode));
        is(menuitem.defaultKeyBinding, (MAC ? "⌃l" : "Ctrl+l"),
           "Wrong keyboard shortcut on " + prettyName(this.menuitemNode));
      }

      this.getID = function openMenu_getID()
      {
        return "menuitem accesskey and shortcut test " +
          prettyName(this.menuItemNode);
      }
    }

    var gQueue = null;
    function doTest()
    {
      gQueue = new eventQueue();
      gQueue.push(new openMenu("menu", "menuitem"));
      gQueue.invoke(); // Will call SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  ]]>
  </script>

  <hbox flex="1" style="overflow: auto;">
    <body xmlns="http://www.w3.org/1999/xhtml">
      <a target="_blank"
         href="https://bugzilla.mozilla.org/show_bug.cgi?id=672092"
         title="Reorganize access key and keyboard shortcut handling code">
        Mozilla Bug 672092
      </a><br/>
      <p id="display"></p>
      <div id="content" style="display: none">
      </div>
      <pre id="test">
      </pre>
    </body>

    <vbox flex="1">
      <keyset>
        <key key="l" modifiers="control" id="key1"/>
      </keyset>

      <menubar>
        <menu label="menu" id="menu" accesskey="u">
          <menupopup>
            <menuitem accesskey="p" key="key1" label="item1" id="menuitem"/>
          </menupopup>
        </menu>
      </menubar>

      <vbox id="debug"/>
    </vbox>
  </hbox>

</window>
